Implemented Physicians Delete Functionalities. Closes #23
Refactor and add deletion methods in MainViewModel

Renamed the `Delete` method in `MainViewModel` to `DeletePatient` for clarity. Added a new `DeletePhysician` method to handle physician deletion and notify changes to the `Physicians` collection.

Updated `MainPage.xaml.cs` to use the renamed `DeletePatient` method in the `DeletePatientClicked` event handler. Added support for the new `DeletePhysician` method in the `DeletePhysiciansClicked` event handler.