Make idCounter private and fix typo in error message
Changed idCounter fields in Patient and Physician classes from
public to private to improve encapsulation. Removed the default
constructor in the Patient class to enforce the use of the
parameterized constructor. Updated Id assignment in the
Physician constructor to explicitly reference the instance
property. Corrected a typo in PhysicianView.xaml.cs to display
"Physician" instead of "Patient" in an error message.